Revert "Infer policy name when given an Array"
This doesn't quite work becuase there's ambiguity when an empty Array is
given. For now we'll have to continue to explicitly define the policy
when passing in Arrays.

I'm thinking that to implement this functionality the right way, we'd
have a `NullPolicy` that we'd fall back to when given an empty array or
nil. It would simply return nil or an empty array for all `Filter` and
`Authorization` methods, respectively. `method_missing` would be used to
work with all public methods.

This reverts commit 9adf566d57d848ad2de3648f293a5049f092f7c1.