Allow :except option to permit_all matchers
In quite a few situations, I've found myself wanting to say that a user
can do everything except one or two. In many of our policies, we've
accumulated a lot of special-purpose filters with long names, and it's
cumbersome to type them all out just to say a user can't :delete.